% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complaints}
\alias{complaints}
\title{Complaints}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\source{
Wayne Taylor (variation.com)
}
\usage{
complaints
}
\description{
A data frame with 20 rows and 3 variables: \itemize{ \item month Month. 
\item  Number of complaints. \item sales Sales volume. }
}
\references{
\url{https://variation.com/normalized-individuals-control-chart/}
}
\keyword{datasets}
