% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospital_infections}
\alias{hospital_infections}
\title{Hospital infections}
\format{A data frame with 432 rows and 5 variables: \itemize{ \item
  \strong{hospital} Abbreviated hospital name. \item \strong{infection} Type
  of infection. BAC: Bacteremia, CDI: Clostridium difficile infection, UTI,
  Urinary tract infection. \item \strong{month} First day of month. \item
  \strong{n} Number of cases. \item \strong{days} Number of risk days. }}
\source{
\url{http://www.haiba.dk/}
}
\usage{
hospital_infections
}
\description{
A dataset containing the number and types of hospital acquired infections in
six hospitals in the Capital Region of Denmark 2015-2016.
}
\keyword{datasets}
