% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lots}
\alias{lots}
\title{Lots}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\source{
Wayne Taylor (variation.com)
}
\usage{
lots
}
\description{
A data frame with 20 rows and 3 variables: \itemize{ \item lot Lot number. 
\item value Average value. \item n Lot size. }
}
\references{
\url{https://variation.com/normalized-individuals-control-chart/}
}
\keyword{datasets}
