% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhs_accidents}
\alias{nhs_accidents}
\title{NHS accidents}
\format{A data frame with 20 rows and 3 variables: \itemize{ \item i Week 
  number. \item r Attendances seen within 4 hours. \item n Total number of 
  attendances. }}
\source{
Mohammed MA, et al. Quality and Safety in Health Care
  2013;22:362–368. \url{https://doi.org/10.1136/bmjqs-2012-001373}.
}
\usage{
nhs_accidents
}
\description{
The number of attendances to major accident and emergency hospital 
departments in the NHS that were seen within 4 hours of arrival over twenty 
weeks.
}
\keyword{datasets}
