% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qicharts2.R
\name{qicharts2}
\alias{qicharts2}
\alias{qicharts2-package}
\title{\code{qicharts2} package}
\description{
Statistical Process Control chart for R
}
\details{
#' See the README on
\url{https://github.com/anhoej/qicharts2/}
}
