% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cabg}
\alias{cabg}
\title{CABG}
\format{A data frame with 2205 rows and 7 variables: \itemize{ \item \strong{no}
  Operation number. \item \strong{data} Date of operation. \item \strong{age}
  Patient age in years. \item \strong{gender} Patient gender. \item
  \strong{los} Length og stay in days. \item \strong{death} TRUE if patient
  died with 30 days after surgery. \item \strong{readmission} TRUE if patient
  were readmitted for any reason within 30 days after surgery. }}
\usage{
cabg
}
\description{
A dataset with data on individual coronary artery bypass graft operations.
}
\keyword{datasets}
