% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cabg}
\alias{cabg}
\title{Coronary artery bypass graft operations}
\format{A data frame with 2205 rows and 6 variables: \itemize{ \item
 {data} Date of operation. \item{age} Patient age in years.
  \item{gender} Patient gender. \item{los} Length og stay in
  days. \item{death} TRUE if patient died within 30 days after
  surgery. \item{readmission} TRUE if patient were readmitted for any
  reason within 30 days after surgery. }}
\source{
Omitted for privacy concerns.
}
\usage{
cabg
}
\description{
A dataset with data on individual coronary artery bypass graft operations.
}
\keyword{datasets}
