% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdi}
\alias{cdi}
\title{Clostridium difficile infections}
\format{A data frame with 36 rows and 5 variables: \itemize{ \item 
 {month} Month of observation. \item{n} Number of hospital 
  acquired CDI. \item{days} Number of risk days. A risk day is a 
  patient day without CDI. \item{period} Factor indicating the period
  'pre' or 'post' intervention. \item{notes} Annotations. }}
\source{
\url{http://www.haiba.dk/} (Amager Hvidovre Hospital).
}
\usage{
cdi
}
\description{
A dataset with data on hospital acquired Clostridium difficile infections 
(CDI) before and after an intervention to reduce the risk of CDI.
}
\keyword{datasets}
